delta changes according to sun

+++ chapman_single/chapman_single_Global.h
@@ -13,7 +13,7 @@
 /*        R. Sander, Max-Planck Institute for Chemistry, Mainz, Germany */
 /*                                                                  */
 /* File                 : chapman_single_Global.h                   */
-/* Time                 : Thu Jan 27 11:37:07 2022                  */
+/* Time                 : Thu Jan 27 12:39:23 2022                  */
 /* Working directory    : /home/kyriacos/CyprusInstitute/kpp/chapman_single */
 /* Equation file        : chapman_single.kpp                        */
 /* Output root filename : chapman_single                            */
@@ -30,7 +30,7 @@ extern float C[NSPEC];                          /* Concentration of all species 
   extern float * FIX;
 extern float E[NSPEC];                          /* Error of all species */
 extern float RCONST[NREACT];                    /* Rate constants (global) */
-extern float TIME;                              /* Current integration time */
+extern double TIME;                             /* Current integration time */
 extern float SUN;                               /* Sunlight intensity between [0,1] */
 extern float TEMP;                              /* Temperature */
 extern float RTOLS;                             /* (scalar) Relative tolerance */
